Update W-Source Framework (19-1-2018)

+++ core/database/ConnectDatabaseInfo.php
@@ -83,7 +83,7 @@ abstract class ConnectDatabaseInfo
      */
     protected final static function getPort()
     {
-        return self::$PORT;
+        return (int)self::$PORT;
     }
 
     /**
